enable deploy

+++ src/app/api/writeMdFile/route.ts
@@ -6,13 +6,23 @@ export async function POST(req: NextRequest) {
   const data = await req.json();
   const pdfText = data.pdfText;
 
-  const filePath = path.join(process.cwd(), "tmp/new.md");
+  const filePath = path.join("/tmp", "new.md");
+  const dirPath = path.dirname(filePath);
+
+  if (!fs.existsSync(dirPath)) {
+    return NextResponse.json(
+      { message: `${dirPath} not found` },
+      { status: 500 }
+    );
+  }
 
   try {
     fs.writeFileSync(filePath, pdfText);
   } catch (error) {
-    console.error(error);
-    return NextResponse.json({ error: "file write error" }, { status: 500 });
+    return NextResponse.json(
+      { error: `file not found in ${filePath}` },
+      { status: 500 }
+    );
   }
 
   return NextResponse.json(
